#32 - Criado metodo DELETE da tasklist.

+++ server/core/views.py
@@ -8,22 +8,23 @@ from django.views.decorators.http import require_http_methods
 from django.views.decorators.csrf import csrf_exempt
 from .models import Tasklist, Task
 
-# Create your views here.
 
 @require_http_methods(["GET","POST"]) 
 def tasklists(request):
     try: 
         tasks = Tasklist.objects.all().values()
     except Tasklist.DoesNotExist: 
-        return JsonResponse({'message': 'Tasklists do not exist'}, status=status.HTTP_404_NOT_FOUND) 
+        return JsonResponse({'message': 'Tasklists do not exist'}, status=status.HTTP_404_NOT_FOUND)
     
     if request.method == "GET":
         task_list = list(tasks)  
         return JsonResponse(task_list, safe=False)
     else:
         return HttpResponse("POST")    
+
+
 @csrf_exempt
-@require_http_methods(["PUT","GET"])
+@require_http_methods(["PUT", "GET", "DELETE"])
 def tasklistsid(request, tasklist_id):
     try: 
         tasklist = Tasklist.objects.get(id=tasklist_id)     
@@ -32,6 +33,7 @@ def tasklistsid(request, tasklist_id):
 
     if request.method == "GET":
         return JsonResponse(model_to_dict(tasklist), safe=False)
+
     elif request.method == "PUT" or request.method == "PATCH":
         
         data = json.loads(request.body)
@@ -42,17 +44,10 @@ def tasklistsid(request, tasklist_id):
             return HttpResponse('ERRO - Id é invalido', status=404) 
        
         return HttpResponse('OK', status=200)
-
-
-@csrf_exempt
-@require_http_methods(["DELETE"])
-def tasklist_delete(request, tasklist_id):
-    tasklist = Tasklist.objects.get(id=tasklist_id)
-    try:
+    
+    elif request.method == "DELETE":
         tasklist.delete()
-        return JsonResponse({"message": 'Task deleted'}, status=200)
-    except Tasklist.DoesNotExist:
-        return JsonResponse({'message': 'Tasklist does not exist!'}, status=404)
+        return JsonResponse({'mesage': 'Tasklist foi deletada!'}, status=204)
 
 
 #TASKS
@@ -66,7 +61,7 @@ def tasklists_id_tasks(request, tasklist_id):
 # GET /tasklists/:id/ - retorna apenas uma tasklist ✅ 
 # POST /tasklists/ - cria uma nova tasklist 
 # PUT/PATCH /tasklists/:id/ - atualiza apenas uma tasklist ✅
-# DELETE /tasklists/:id/ - deleta apenas uma tasklist
+# DELETE /tasklists/:id/ - deleta apenas uma tasklist ✅
 
 # Tasks
 # GET /tasklists/:id/tasks/ - retorna todas as tasks de uma determinada tasklist ✅
